refactor(checker): split comparability query boundary
